applications: asset_tracker_v2: Add LwM2M codec helper library
Add LwM2M codec helper library and convert LwM2M codec test to test
the new helper library. A separate test for the LwM2M codec will be
added later.

The LwM2M codec helpers library was added to improve:
 - Readability in the LwM2M codec backend.
 - Testability. The new helpers library comprises of smaller APIs with
   specific purposes that in turn makes each test case smaller and
   more understandable.

Fixes CIA-739, CIA-647
