+++ applications/asset_tracker_v2/overlay-lwm2m.conf
@@ -65,19 +65,13 @@ CONFIG_COAP_EXTENDED_OPTIONS_LEN_VALUE=40
 
 # IPSO v1.1
 CONFIG_LWM2M_IPSO_SUPPORT=y
-CONFIG_LWM2M_IPSO_TEMP_SENSOR_VERSION_1_1=y
-CONFIG_LWM2M_IPSO_HUMIDITY_SENSOR_VERSION_1_1=y
-CONFIG_LWM2M_IPSO_PRESSURE_SENSOR_VERSION_1_1=y
 CONFIG_LWM2M_IPSO_PUSH_BUTTON_VERSION_1_1=y
 CONFIG_LWM2M_CONNMON_OBJECT_VERSION_1_2=y
 
 # Zephyr LwM2M standard objects
-CONFIG_LWM2M_IPSO_PRESSURE_SENSOR=y
 CONFIG_LWM2M_IPSO_PUSH_BUTTON=y
 CONFIG_LWM2M_CONN_MON_OBJ_SUPPORT=y
 CONFIG_LWM2M_LOCATION_OBJ_SUPPORT=y
-CONFIG_LWM2M_IPSO_HUMIDITY_SENSOR=y
-CONFIG_LWM2M_IPSO_TEMP_SENSOR=y
 CONFIG_LWM2M_CONN_MON_BEARER_MAX=2
 
 # Use nrf-sdk lwm2m client utils for convenience. The library contains pre-existing support
