+++ applications/asset_tracker_v2/tests/lwm2m_codec_helpers/Kconfig
